Refactor into "MVC"
This commit refactors the whole code to have a more modular, "MVC" style
architecture.

There's a wave model that represents the wave data. Separately from that
there's a wave view which handles the rendering and updating.

You can now also modify a wave by drawing directly in the canvas. This
will automatically update the table data as well as the canvas.