Smooth out drawing
There was a problem with the mouse drawing because the mouse move event
skips pixels if you drag too fast. This was leaving gaps in the graph
and data.

This commit fixes that by filling in the missing gaps by interpolating
(linearly) the data from now to the previous point (if the gap is
greater than 1, that is to say, if we skipped a pixel).